+++ app/workers/cron_worker/scores_recalculation.rb
@@ -1,6 +1,6 @@
 class CronWorker::ScoresRecalculation
   def self.perform
-    WriteLog.info(self, 'started')
+    Zazo::Tools::Logger.info(self, 'started')
     Contact.expired.each do |contact|
       contact.scores.destroy_all
       Score::CalculationByContact.new(contact).do